sagit: configs: nfc: enable pll clock manipulation
The dts specifies the existance of a pll clock that needs to be
manipulated. Keeping this clock on during suspend causes battery drain,
and because the config specifies the existance of a crystal-based clock,
the clocks are never turned off.

This is a theory. I don't know who is wrong, the dts or the config.
I also don't know why it works fine with CAF.

Change-Id: I8c6be9e9d0382fa040a067e7848b9851fa939c63